add fix manual conpilation and add nccl

+++ 3_install_cuda.sh
@@ -52,14 +52,28 @@ cudnn() {
     read -n1 -r -p "Install libcupti-dev. press ENTER to continue!" ENTER
     sudo apt install libcupti-dev -y
 
-    echo "[CUDA-TSFLOW] Backup linux enviroment before add cuda enviroment.."
-    cp ~/.bashrc ~/.bashrc.backup.cuda
-
     echo "[CUDA-TSFLOW] Cofiguring cuda in linux enviroment.."
-    echo " " >> ~/.bashrc
-    echo '# CUDA Enviroment' >> ~/.bashrc
-    echo 'export LD_LIBRARY_PATH="$LD_LIBRARY_PATH:/usr/local/cuda/lib64:/usr/local/cuda/extras/CUPTI/lib64"' >> ~/.bashrc
-    echo 'export CUDA_HOME="/usr/local/cuda"' >> ~/.bashrc
+    read -p "Have you set the cuda in ~./bashrc? [y/N]: "
+    UPDATE="${UPDATE:=N}"
+    if [[ $UPDATE = "N" ]] || [[ $UPDATE = "n" ]]; then
+      nccl
+    elif [[ $UPDATE = "Y" ]] || [[ $UPDATE = "y" ]]; then
+      echo "[CUDA-TSFLOW] Backup linux enviroment before add cuda enviroment.."
+      cp ~/.bashrc ~/.bashrc.backup.cuda
+
+      echo "[CUDA-TSFLOW] Cofiguring cuda in linux enviroment.."
+      echo " " >> ~/.bashrc
+      echo '# CUDA Enviroment' >> ~/.bashrc
+      echo 'export LD_LIBRARY_PATH="$LD_LIBRARY_PATH:/usr/local/cuda/lib64:/usr/local/cuda/extras/CUPTI/lib64"' >> ~/.bashrc
+      echo 'export CUDA_HOME="/usr/local/cuda"' >> ~/.bashrc
+      echo 'export PATH="$PATH:$HOME/Bazel/bin"' >> ~/.bashrc
+
+      nccl
+    else
+      echo "[CUDA-TSFLOW] Input invalid."
+      echo "[CUDA-TSFLOW] Installation aborted."
+      exit
+    fi
   else
     echo "[CUDA-TSFLOW] CUDNN installer not found."
     echo "[CUDA-TSFLOW] CUDNN installation failed"
@@ -67,6 +81,20 @@ cudnn() {
   fi
 }
 
+nccl() {
+  read -n1 -r -p "Install NCCL. press ENTER to continue!" ENTER
+  if [[ -f installer/${NCCL} ]]; then
+    echo "[CUDA-TSFLOW] Installing NCCL.."
+    mkdir installer/nccl
+    tar Jxvf installer/${NCCL} -C installer/nccl/ --strip-components=1
+    sudo cp installer/nccl/lib/libnccl* /usr/local/cuda/lib64/
+    sudo cp installer/nccl/include/nccl.h /usr/local/cuda/include/
+  else
+    echo "[CUDA-TSFLOW] NCCL installer not found."
+    echo "[CUDA-TSFLOW] NCCL installation failed"
+    exit
+  fi
+}
 
 echo "==============================================================================================="
 echo "Welcome to Cuda Installer"
@@ -114,10 +142,12 @@ if [[ -L /usr/local/cuda ]]; then
   exit
 fi
 
-read -p "Please specify CUDA installer file. [Defaut: cuda_9.0.176_384.81_linux.run]:" CUDA
+read -p "Please specify CUDA installer file. [Defaut: cuda_9.0.176_384.81_linux.run]: " CUDA
 CUDA="${CUDA:=cuda_9.0.176_384.81_linux.run}"
-read -p "Please specify CDNN installer file. [Defaut: cudnn-9.0-linux-x64-v7.4.2.24.tgz]:" CUDNN
+read -p "Please specify CDNN installer file. [Defaut: cudnn-9.0-linux-x64-v7.4.2.24.tgz]: " CUDNN
 CUDNN="${CUDNN:=cudnn-9.0-linux-x64-v7.4.2.24.tgz}"
+read -p "Please specify NCCL installer file. [Defaut: cudnn-9.0-linux-x64-v7.4.2.24.tgz]: " NCCL
+NCCL="${NCCL:=nccl_2.3.7-1+cuda9.0_x86_64.txz}"
 echo " "
 
 cuda
